+++ SM/Application/Features/Products/Commands/UpdateProduct/UpdateProductCommand.cs
@@ -12,7 +12,7 @@ namespace Application.Features.Products.Commands.UpdateProduct
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Rate { get; set; }
+        public int Rate { get; set; }
         public class UpdateProductCommandHandler : IRequestHandler<UpdateProductCommand, Response<int>>
         {
             private readonly IProductRepositoryAsync _productRepository;
